refactor: split media upgrade pipeline
